[receiver/solacereceiver] move model to internal package
