+++ receiver/solacereceiver/internal/model/README.md
