+++ receiver/solacereceiver/internal/model/egress_v1.proto
