+++ receiver/solacereceiver/internal/model/receive_v1.proto
